Announcement dropdown
The announcement dropdown is created. The announcement dropdown is made with wtforms, where a queryselectfield was used to pass the values of the titles to the select field. Then, on submit, the data from the forms was turned into the string, and all characters that were not numbers were removed, such that id was the only one left. The id was then used to get current_announcement to then be displayed.